fix(api): make runtime_config administrator-only
Serving the runtime config to every authenticated user relied on a denylist:
`_redact_config_secrets` masks the credential fields we know about, so any
future secret added to `InvokeAIAppConfig` would leak by default until someone
remembered to update it. Admin-only is an allowlist posture without that
failure mode.

Nothing outside the admin UI actually needed it. The three non-admin call
sites read `config.multiuser` purely to derive "admin or single-user", which is
already available from the unauthenticated `/auth/status` route as
`multiuser_enabled` — the same source `useIsModelManagerEnabled` uses. The
remaining fields (`max_queue_history`, `image_subfolder_strategy`) only feed
admin-gated edit controls.

- Extract that predicate into `useIsAdmin`, and have `useIsModelManagerEnabled`
  delegate to it. Semantics are preserved exactly, so this is a pure refactor.
- Use it in the three settings components instead of `runtime_config`, and skip
  the query for non-admins so they generate no failed requests.
- AboutModal's debug blob now omits the config section for non-admins. Its
  client-side redaction is dropped: it only ever masked `remote_api_tokens` and
  never the `external_*_api_key` fields, so a logged-in non-admin could read
  provider API keys out of it. The server now masks both.

Server-side masking is kept as defense in depth — an admin's browser has no use
for the raw values either.
