+++ invokeai/frontend/web/src/features/auth/hooks/useIsAdmin.ts
@@ -1,0 +1,29 @@
+import { useAppSelector } from 'app/store/storeHooks';
+import { selectCurrentUser } from 'features/auth/store/authSlice';
+import { useMemo } from 'react';
+import { useGetSetupStatusQuery } from 'services/api/endpoints/auth';
+
+/**
+ * Hook to determine whether the current user holds administrator privileges.
+ *
+ * Returns true if:
+ * - Multiuser mode is disabled (single-user mode = always admin)
+ * - Multiuser mode is enabled AND the user is an admin
+ *
+ * This mirrors the backend's `AdminUserOrDefault` dependency. Prefer it over reading `multiuser` from
+ * `runtime_config`, which is itself an admin-only route.
+ */
+export const useIsAdmin = (): boolean => {
+  const user = useAppSelector(selectCurrentUser);
+  const { data: setupStatus } = useGetSetupStatusQuery();
+
+  return useMemo(() => {
+    // If multiuser is disabled, treat as admin (single-user mode)
+    if (setupStatus && !setupStatus.multiuser_enabled) {
+      return true;
+    }
+
+    // If multiuser is enabled, check if user is admin
+    return user?.is_admin ?? false;
+  }, [setupStatus, user]);
+};
