+++ invokeai/frontend/web/src/features/system/components/SettingsModal/SettingsModal.tsx
@@ -21,7 +21,7 @@ import { useAppDispatch, useAppSelector } from 'app/store/storeHooks';
 import { InformationalPopover } from 'common/components/InformationalPopover/InformationalPopover';
 import ScrollableContent from 'common/components/OverlayScrollbars/ScrollableContent';
 import { buildUseBoolean } from 'common/hooks/useBoolean';
-import { selectCurrentUser } from 'features/auth/store/authSlice';
+import { useIsAdmin } from 'features/auth/hooks/useIsAdmin';
 import { selectShouldUseCPUNoise, shouldUseCpuNoiseChanged } from 'features/controlLayers/store/paramsSlice';
 import { ExternalProviderStatusList } from 'features/system/components/SettingsModal/ExternalProviderStatusList';
 import { useRefreshAfterResetModal } from 'features/system/components/SettingsModal/RefreshAfterResetModal';
@@ -88,8 +88,9 @@ const SettingsModal = (props: { children: ReactElement<{ onClick?: () => void }>
 
   const settingsModal = useSettingsModal();
   const refreshModal = useRefreshAfterResetModal();
-  const currentUser = useAppSelector(selectCurrentUser);
-  const { data: runtimeConfig } = useGetRuntimeConfigQuery();
+  const canEditRuntimeConfig = useIsAdmin();
+  // runtime_config is an admin-only route; don't fire it for non-admins just to render disabled controls.
+  const { data: runtimeConfig } = useGetRuntimeConfigQuery(undefined, { skip: !canEditRuntimeConfig });
   const [updateRuntimeConfig, { isLoading: isUpdatingRuntimeConfig }] = useUpdateRuntimeConfigMutation();
   const pendingMaxQueueHistoryRef = useRef<number | null | undefined>(undefined);
 
@@ -107,7 +108,6 @@ const SettingsModal = (props: { children: ReactElement<{ onClick?: () => void }>
   const shouldConfirmOnNewSession = useAppSelector(selectSystemShouldConfirmOnNewSession);
   const shouldShowInvocationProgressDetail = useAppSelector(selectSystemShouldShowInvocationProgressDetail);
   const maxQueueHistory = runtimeConfig?.config.max_queue_history ?? null;
-  const canEditRuntimeConfig = runtimeConfig ? !runtimeConfig.config.multiuser || currentUser?.is_admin : false;
   const [maxQueueHistoryInputState, setMaxQueueHistoryInputState] = useState(() => ({
     source: maxQueueHistory,
     value: formatOptionalInteger(maxQueueHistory),
